fixing to launch app

+++ composeApp/src/commonMain/kotlin/org/softwareanvil/domain/generator/NameGenerator.kt
@@ -33,5 +33,5 @@ class NameGenerator(
     }
 
     private fun pick(list: List<Syllable>): Syllable =
-        weightedRandom(list, { it.weight }, random)
+        weightedRandom(list, { it.weight.toInt() }, random)
 }
